+++ src/components/modals/CreateCollectionModal.vue
@@ -28,13 +28,13 @@
             id="collectionName"
             v-model="newCollection.name"
             type="text"
-            class="vdb-c-w-full vdb-c-px-10 vdb-c-py-8 vdb-c-text-[14px] vdb-c-border vdb-c-border-gray-300 vdb-c-rounded-md vdb-c-box-border focus:vdb-c-border-orange-500 focus:vdb-c-outline-none"
+            class="vdb-c-w-full vdb-c-px-10 vdb-c-py-8 vdb-c-text-[14px] vdb-c-text-kilvish-900 vdb-c-border vdb-c-border-gray-300 vdb-c-rounded-md vdb-c-box-border focus:vdb-c-border-orange-500 focus:vdb-c-outline-none"
             placeholder="Collection Name (required)"
           />
           <textarea
             id="collectionDescription"
             v-model="newCollection.description"
-            class="vdb-c-w-full vdb-c-px-10 vdb-c-py-8 vdb-c-text-[14px] vdb-c-border vdb-c-border-gray-300 vdb-c-rounded-md vdb-c-box-border focus:vdb-c-border-orange-500 focus:vdb-c-outline-none"
+            class="vdb-c-w-full vdb-c-px-10 vdb-c-py-8 vdb-c-text-[14px] vdb-c-text-kilvish-900 vdb-c-border vdb-c-border-gray-300 vdb-c-rounded-md vdb-c-box-border focus:vdb-c-border-orange-500 focus:vdb-c-outline-none"
             placeholder="Description"
           ></textarea>
         </div>